Implementado createStoreMap
-- Implementado uma função de primeira classe no módulo de stores que irá retornar uma "getStore", esse método recebe a classe que será feito o map e cuidará de todo o processo de instanciação e retorno de classes que são necessários ter várias instancias
-- O getStore da Depth foi substituído pelo createStoreMap
