+++ client/src/routes/market/_stores/depth.ts
@@ -1,4 +1,4 @@
-import Store from '../../../utils/store'
+import Store, { createStoreMap } from '../../../utils/store'
 import { addSocketListener } from '../../../utils/websocket'
 import type { MarketDepth } from '../depth'
 import type { SuportedCurrencies } from '../../../../../src/libs/currencies'
@@ -22,23 +22,4 @@ class DepthStore extends Store<MarketDepth[]> {
 	}
 }
 
-const map = new Map<string, DepthStore>()
-
-/**
- * Retorna uma DepthStore das currencies requisitadas
- * @param base A currency Base desse par
- * @param target A currency Target desse par
- */
-export default function getDepthStore(
-	base: SuportedCurrencies,
-	target: SuportedCurrencies,
-) {
-	if (base == target) throw new Error('Currency base must be different from Currency target')
-	const mapKey = [base, target].join(',')
-	let store = map.get(mapKey)
-	if (!store) {
-		store = new DepthStore(base, target)
-		map.set(mapKey, store)
-	}
-	return store
-}
+export const getDepthStore = createStoreMap(DepthStore)
